Initial commit
The manifests are coming mostly from
https://github.com/poseidon/terraform-render-bootstrap/tree/v0.14.0/resources/manifests,
except that args for daemons are changed and reduced to the minimum.

There is also plenty of IP addresses hardcoded around, as this is
initial import of my work.

All the values specified in values.yaml file should be filled in with
valid certificates. The code to generate those certificates will be
shared later.

Signed-off-by: Mateusz Gozdek <mgozdekof@gmail.com>